+++ db/migrate/20160623203435_update_chapters_slug_index.rb
@@ -1,0 +1,11 @@
+class UpdateChaptersSlugIndex < ActiveRecord::Migration
+  def up
+    remove_index :chapters, :slug
+    add_index :chapters, :slug, unique: true
+  end
+
+  def down
+    remove_index :chapters, :slug
+    add_index :chapters, :slug
+  end
+end